Added import of sparse content-class definitions
These definitions define the classes that are required by the
rest of the object import, the system then verifies that existing
classes has the attributes and types as defined. Existing classes
may have more attributes but the ones that are referenced needs
to match.

Remapping of class identifier and attribute identifiers are
possible.