Support for changing existing class attributes
setAttribute() is specified with an identifer and it requires that the
attribute exists.

If the type is changed a callback function may be used to transform
all existing attributes to new values, specify the 'objectTransform'
field.

isRequired, isSearchable, isInformationCollector and canTranslate are
now null by default, this means that they only update the attribute if
set to a non-null value. For creation they will still create with
default values.