Added support for transformation of content object data
It consists of several systems.
- Mapping of object uuid to a new value
- Mapping of node uuid to a new value, any location parents are also
  remapped with this.
- Transformation classes to run objects, either by uuid, class or
  in general.

Original uuids are kept in separate fields, prefixed with 'original_'.

Transformation classes may return new data for the objects, for instance
checking for existing objects by another attribute.
They may also set the 'update_attributes' entry which determines
which attributes may be written back, this is useful for existing
objects which have received new attributes.
Additionally it may set 'action_update_object' to an array of object
data to be updated, this either 'object' for main data for object,
'attribute' for attributes and 'location' for writing all location
nodes.

The exising parent remapping was changed into the generic node
remapping.

Slight change in transform classes, the imported instance is now
passed as the first parameter to the constructor, the second is the
ini file.