Fixed issue with importing the same object multiple times
This can happen when importing multiple files, now it will create the
object once but locations will be merged if there are new locations.
This also fixes the issue where an object has no locations but later
on gets a proper import with locations.
Also fixed setting up 'locations' on object which defines uuid, parent,
and also optionally priority and visibility.