Added a batch processor class for easily modifying multiple objects
It takes a QuerySet object as parameter (or parameters for QuerySet)
and takes care of iterating over all objects in that set and calling
a callback. This class can either be used directly by passing
functions into the callbacks or by subclassing.
The processor defaults to iterating over objects (ie. main nodes) and
will pass a ContentObject instance to the callback, however it may be
changed to visiting nodes instead.
There are also callbacks for when an object was modified, skipped or
one for all visited nodes/objects.