Extended BatchProcessor with match control, visit callback and read-only support
The 'match' callback is called for each node and if it returns true
then the node is processed, if false it ignores the node. This can
be used for more advanced filtering.
The 'visit' callback is called before node/object is processed, can
be used to output progress before anything happens.
By setting processor param 'readOnly' to true it will force the
processor to no call the 'process' callback, 'visit' and 'visited'
are still called.