Support for late remapping of object relations
If a relation (owner, related object, embed) refers to a uuid that is
not yet added it records the relationship. When the referred object
is added later on and chanages the uuid, e.g. due to a remappinm, it
will then find all other objects that relates to it and fix the uuid
that they refer to.