Improved access to class attribute values
The `value` property on ContentTypeAttribute is now a dynamic
property. It will either fetch the value from the database and
decode it into export/import format, or for new/changed attributes
it returns the value set in memory.
A new property `isStored` is set to true for attribute objects
that are stored in the database, false if the value is in memory
only.